pyproject.toml
[project]
name = "mcp-inandout"
version = "0.1.0"
description = "Starter project for experimenting with MCP client and server integrations"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]
